Remove Interface Soe Search Indexing Buffer
