+++ sdk/search/search-documents/test/internal/geographyPoint.spec.ts
@@ -11,7 +11,7 @@ describe("geographyPoint", () => {
     const result = JSON.parse(JSON.stringify(geoPoint));
     assert.deepEqual(result, {
       type: "Point",
-      coordinates: [-122.123889, 47.669444],
+      coordinates: [47.669444, -122.123889],
       crs: { type: "name", properties: { name: "EPSG:4326" } }
     });
   });
